app/requirements.txt
alembic==1.11.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
gensim==4.3.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
Mako==1.2.4
MarkupSafe==2.1.2
nltk==3.8.1
numpy<=1.25.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0
scipy<=1.11.2
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.14
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.5.0
urllib3==2.0.4
Werkzeug==2.3.4
transformers==4.32.1
sentencepiece==0.1.99
gunicorn==21.2.0
pandas<=2.1.0
langchain==0.0.300
wikipedia==1.4.0
spacy==3.7.1
chromadb==0.4.12
